Put `entries` and `load` in same file
Feels much tider
